feat: Delete all pycache files (oops) and update readme
The readme contains hopefully full instructions, as far as I have right
now, for how to install and test with the package, especially as a
contributor. Also has a writeup of all the settings and how you can use
them.

I also realised I had added all the pycache files and didn't have a
proper gitignore so hopefully this will fix that one xD