feat: Add show_warning_if_found gdpr meta option
This is a very basic implementation that does not allow for custom error
messages, I just wanted to get it working so we can roll this out asap.
We want to show these warnings so its very easy to tell if a GDPR
request is going to be simple or not. Basic rules being if something is
found in a specific table it will then say to contact a the person
handling the GDPR request to discuss further as we have a number of
protocals we need to follow if a GDPR request is non standard (i.e. they
did more than just sign up)