test: Models, add in migrations, remove test app
I went to test models and it errored almost immediatly as it was missing
migrations required to run the tests. And thus the journey began to add
migrations to the test app.

I removed the test_app entirely because I couldn't import it as an app
if it was nested (or atleast didn't look into it too much) and I
    realised using the actual "test" app would be easier with less
    setup.

The model tests are not fully in I have only written the first one, but
hey it works!

Also added a migrations folder to the GDPR manager app. Right now we
have no migrations but if for some reason in future we need one its
pretty confusing when no migrations generate because the migraitons
folder is missing and doesn't have a __init__.py.