feat: Add GDPR_MANAGER_EXCLUDE_MODELS to handle join tables
When you have a many to many join in django it creates a join table, its
just a list of ID's and has no reason to have a PII in it so they need a
way to be excluded.

I did considering adding them into the GDDPR_MANAGER_EXCLUDE list but
frankly it seemed like making things weird because I didn't like my
original naming. Which is a bit of a "oh tough shit" kinda situation.