fix: add dummy migration for GDPRManager model
This is a model that's only used for the admin, and doesn't actually
reach the database at all.

But we need a dummy migration in order to prevent Django from deteching
"changes" in the model files.

Adding this should make it so Django projects using this module don't
detect model changes when they shouldn't.