test: Update test that checks empty value searches
I could have created another test for this but it seemed a little like
overdoing it when I could just switch around the search order in the
empty value search and it tests what I need it to properly. As far as
I'm concerned its just making sure this test works properly anyway.